Chore: rename folder studies to study
